Update namings, introduce interactions.
